* '4.7' of scm.dev.nokia.troll.no:qt/oslo-staging-2:
  Fix warnings in the declarative module
  compile fix for mingw (also removes some warnings)
  Sort indicators displayed incorrectly in GTK style

+++ src/gui/kernel/qapplication_win.cpp
@@ -2278,7 +2278,7 @@ extern "C" LRESULT QT_WIN_CALLBACK QtWndProc(HWND hwnd, UINT message, WPARAM wPa
         case WM_GETOBJECT:
             {
                 // Ignoring all requests while starting up
-                if (QApplication::startingUp() || QApplication::closingDown() || (DWORD)lParam != OBJID_CLIENT) {
+                if (QApplication::startingUp() || QApplication::closingDown() || (LONG)lParam != OBJID_CLIENT) {
                     result = false;
                     break;
                 }
